+++ Integrations/MicrosoftGraphMail/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
-
+- added ***msgraph-mail-list-folders*** command.
+- added ***msgraph-mail-list-child-folders*** command.
+- added ***msgraph-mail-create-folder*** command.
+- added ***msgraph-mail-update-folder*** command.
+- added ***msgraph-mail-delete-folder*** command.
+- added ***msgraph-mail-move-email*** command.
+- added ***msgraph-mail-get-email-as-eml*** command.
 
 ## [19.10.1] - 2019-10-15
   - Improved the description of the *search* argument in ***msgraph-mail-list-emails*** command.